+++ fusion-endpoint/src/main/java/com/vaadin/fusion/startup/ConnectEndpointsValidator.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.server.startup.fusion;
+package com.vaadin.fusion.startup;
 
 import javax.servlet.ServletContext;
 import javax.servlet.ServletException;
